Add support for running concurrent tasks.
